🐛 fix firefox UUID
The UUID (Universally Unique Identifier) of a Firefox extension or add-on is a identifier used by the browser to identify the specific extension. It is not considered a secret and is used by the browser to load and run the extension.

The UUID is included in the extension's installation package and is visible to anyone who has access to the package. It is not meant to be kept secret and is not used for security purposes. However, it is important to ensure that the extension is obtained from a trusted source, as extensions can potentially access and modify sensitive data in your browser.